Install extensions in a shared schema
This way all tenants will be able to access them instead of just the
default one.

The apartment gem recommends using a rake task instead of a migration,
but that's a solution which is primarily meant for new installations.
Migrations are easier to execute on existing installations.

However, since this migration doesn't affect the `schema.rb` file, we
still need to make sure the shared schema is created in tasks which do
not execute migrations, like `db:schema:load` or `db:test:prepare`, just
like the apartment gem recommends. That's why we're enhancing these
tasks so they execute this migration.

Note we need to drop existing extensions and create them again on a
shared schema, since extensions can only be installed into one schema
per database. If we didn't drop them, the command to create them would
be ignored on existing CONSUL installations, meaning the application
would work normally for the default tenant, but it would crash for other
tenants in places where extensions like `unaccent` are required.

Dropping extensions would be harmful if we used certain extensions, like
the Hstore extension, which would drop all columns using the `hstore`
type, leading to permanent data loss. However, since we only use
extensions providing certain functions like `unaccent`, dropping them
and creating them again is harmless.

This might also work if we had extensions like Hstore, but I haven't
tested it:

```
execute "CREATE EXTENSION IF NOT EXISTS #{extension} SCHEMA #{schema};"
execute "ALTER EXTENSION #{extension} SET SCHEMA #{schema};"
```

Also note the plpgsql is not handled here. This is a special extension
which must be installed on the pg_catalog schema, which is always in the
search path and so is shared by all tenants.

Finally, we also need to change the `database.yml` file in order to
search for shared extensions while running migrations or model tests,
since none of our enabled extensions are executed during migrations;
we're also adding a rake task for existing installations. Quoting the
apartment documentation:

> your database.yml file must mimic what you've set for your default and
> persistent schemas in Apartment. When you run migrations with Rails,
> it won't know about the extensions schema because Apartment isn't
> injected into the default connection, it's done on a per-request
> basis.