Add shared_extensions to the schema search path
When we add multitenancy support, we're going to add a
`shared_extensions` schema. In order to ease the upgrade, we're adding
support for this schema now.

Other than updating the database YAML files which will be used in new
applications, we're providing a rake task so existing installations can
apply these changes too.