Add an interface to manage tenants TODO: test
Note we aren't allowing to delete a tenant because it would delete all
its data, so this action is a very dangerous one. We might need to add a
warning when creating a tenant, indicating the tenant cannot be
destroyed.

For now, we're letting administrators of the "main" (default) tenant to
create other tenants. However, since this is a setting that is a bit
risky to handle when people don't know what multitenancy is, for now
we're only allowing to manage tenants when the multitenancy
configuration option is enabled.

We've thought about creating a new role to manage tenants or a new URL
out of the admin area. We aren't doing so for simplicity purposes and
because we want to keep Consul working the same way it has for
single-tenant installations, but we might change it in the future.
There's also the fact that by default we create one user with a known
password, and if by default we create a new role and a new user to
handle tenants, the chances of people forgetting to change the password
of one of these users increases dramatically, particularly if they
aren't using multitenancy.