Generate one sitemap per tenant
Note that one limitation of the `sitemap:refresh` is that it only pings
search engines at the end, so it only does so for the `default_host`
defined last. So we're using the `sitemap:refresh:no_ping` task instead
and pinging search engines after creating each sitemap.

Note we're pinging search engines in staging and preproduction
environments. I'm leaving it that way because that's what we've done
until now, but I wonder whether that's the right behavior.