Generate one sitemap per tenant TODO: change robots or nginx to map sitemaps
Note that one limitation of the `sitemap:refresh` is that it only pings
search engines at the end, so it only does so for the `default_host`
defined last. So we're using the `sitemap:refresh:no_ping` task instead
and pinging search engines after creating each sitemap.

Also note we're using `ApplicationMailer` to generate a URL in the
sitemap depending on the current tenant. We might change it in the
future to a generic class to generate the URL based on the
`default_url_options`.  But, for now, this is probably good enough.