Allow adding per-environment custom settings
Until now, when editing a specific environment, other Consul
installations had to edit the original file, which made it harder to
upgrade.

Now it's possible to change the default Consul settings using custom
files, making it easier to upgrade to versions of Consul which change
the original environment files (which is very common when upgrading
versions of Rails).