Allow having different SMTP settings per tenant TODO: tests?
Right now this is configured using the `secrets.yml` file, which if the
file we've used in the past to configure SMTP settings.

Note that, in the `current_secrets` method, the `if default?` condition
is there so in single-tenant applications it returns the exact same
object as `Rails.application.secrets`, and it makes it immediately clear
for developers reading the code. We're also caching the subdomain
secrets (using `||=`) so they behave the same way as Rails secrets.

The structure of the SMTP settings in the secrets file should be:

```
production:
  tenants:
    name_of_the_tenant_subdomain:
      smtp_settings:
        address:
        (...)
```