Simplify URL validation in related content
We were using `Setting["url"]` to verify the content belonged to the
application URL, but we can get the URL from the request.

Note that we aren't including the port in the validation because the
port isn't added in the tests when using url helpers like
`polymorphic_url`, so we had some inconsistencies. If this wasn't the
case, we could simply use `root_url` instead of `protocol_with_host`.
IMHO it's a minor thing because production servers rarely include ports
in their URLs (and because the related contents feature isn't so
commonly used).